% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{as.network.Rcpp_DirectedNet}
\alias{as.network.Rcpp_DirectedNet}
\title{convert and DirectedNet to a network object}
\usage{
\method{as.network}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\description{
convert and DirectedNet to a network object
}
