\name{as.network.Rcpp_UndirectedNet}
\alias{as.network.Rcpp_UndirectedNet}
\title{convert and UndirectedNet to a network object}
\usage{
  as.network.Rcpp_UndirectedNet(x, ...)
}
\arguments{
  \item{x}{the object}

  \item{...}{unused}
}
\description{
  convert and UndirectedNet to a network object
}

