% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{class}
\name{BinaryNet}
\alias{BinaryNet}
\alias{DirectedNet}
\alias{UndirectedNet}
\alias{Rcpp_DirectedNet-class}
\alias{Rcpp_UndirectedNet-class}
\title{BinaryNet}
\description{
BinaryNet
}
\details{
Rcpp_DirectedNet and Rcpp_UndirectedNet are the native network classes for the
lolog package. They are designed for algorithmic performance, and are thin wrappers
for an underlying C++ object. These network objects can be passed back and forth between
R and C++ with little overhead. Because they are pointers to C++ objects, serialization
via 'save' or 'dput' are not supported
}
