% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolog.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{calculate model statistics from a formula}
\usage{
calculateStatistics(formula)
}
\arguments{
\item{formula}{An lolog formula}
}
\description{
calculate model statistics from a formula
}

