% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{methods}
\name{call-symbols}
\alias{call-symbols}
\alias{_lolog_initStats}
\alias{_rcpp_module_boot_lolog}
\alias{initLologStatistics}
\alias{runLologCppTests}
\title{Internal Symbols}
\description{
Internal symbols used to access compiles code.
}
