% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolog.R
\name{coef.lolog}
\alias{coef.lolog}
\title{Extracts estimated model coefficients.}
\usage{
\method{coef}{lolog}(object, ...)
}
\arguments{
\item{object}{A `lolog` object.}

\item{...}{unused}
}
\description{
Extracts estimated model coefficients.
}
\examples{
# Extract parameter estimates as a numeric vector:
data(ukFaculty)
fit <- lolog(ukFaculty ~ edges)
coef(fit)
}
