% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolog.R
\name{createCppModel}
\alias{createCppModel}
\title{creates a model}
\usage{
createCppModel(formula, cloneNet = TRUE, theta = NULL,
  modelClass = "Model")
}
\arguments{
\item{formula}{the model formula}

\item{theta}{the model parameters.}

\item{modelClass}{The trailing name of the model class. For now, either Model or ReModel}

\item{ignoreMnar}{ignore missing not at random offsets}
}
\description{
creates a model
}

