% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,Rcpp_DirectedNet-method}
\alias{[,Rcpp_UndirectedNet-method}
\alias{[<-}
\alias{[<-,Rcpp_DirectedNet-method}
\alias{[<-,Rcpp_UndirectedNet-method}
\title{indexing}
\usage{
\S4method{[}{Rcpp_DirectedNet}(x, i, j, ..., maskMissing = TRUE,
  drop = TRUE)

\S4method{[}{Rcpp_UndirectedNet}(x, i, j, ..., maskMissing = TRUE,
  drop = TRUE)

\S4method{[}{Rcpp_DirectedNet}(x, i, j, ...) <- value

\S4method{[}{Rcpp_UndirectedNet}(x, i, j, ...) <- value
}
\arguments{
\item{x}{object}

\item{i}{indices}

\item{j}{indices}

\item{...}{unused}

\item{maskMissing}{should missing values be masked by NA}

\item{drop}{unused}

\item{value}{values to assign}
}
\description{
indexing

indexing

indexing

indexing
}

