% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{gofit}
\alias{gofit}
\title{conduct goodness of fit diagnostics}
\usage{
gofit(object, ...)
}
\arguments{
\item{object}{the object to evaulate}

\item{...}{additional parameters}
}
\description{
conduct goodness of fit diagnostics
}
\details{
see ?gof.lolog
}
