% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{class}
\name{LatentOrderLikelihood}
\alias{DirectedLatentOrderLikelihood}
\alias{LatentOrderLikelihood}
\alias{Rcpp_DirectedLatentOrderLikelihood-class}
\alias{Rcpp_UndirectedLatentOrderLikelihood-class}
\alias{UndirectedLatentOrderLikelihood}
\title{LatentOrderLikelihood}
\description{
LatentOrderLikelihood
}

