% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{lazega}
\alias{lazega}
\title{Collaboration Relationships Among Partners at a New England Law Firm}
\source{
See \url{http://elazega.fr/?page_id=609} and \url{https://www.stats.ox.ac.uk/~snijders/siena/Lazega_lawyers_data.htm}
}
\usage{
data(lazega)
}
\description{
This data set comes from a network study of corporate law partnership that was 
carried out in a Northeastern US corporate law firm, referred to as SG&R, 1988-1991 
in New England.
}
\section{Licenses and Citation}{
 CC BY 4.0. When publishing results obtained using this
data set, the original authors (Lazega, 2001)
should be cited, along with this \code{R} package.
}

\section{Copyright}{

Creative Commons Attribution-Share Alike 4.0 International License, 
see https://creativecommons.org/licenses/by/4.0/ for details.
}

\references{
Lazega, Emmanuel (2001), The Collegial Phenomenon: The Social Mechanisms of Cooperation 
among Peers in a Corporate Law Partnership, Oxford: Oxford University Press
}
\keyword{datasets}
