% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{methods}
\name{lolog-terms}
\alias{lolog-terms}
\title{LOLOG Model Terms}
\description{
LOLOG Model Terms
}
\section{NA}{

\itemize{
\item{\code{boundedDegree(lower,upper)}  (order-independent)  (undirected)}{ 
Adds a constraint that the degrees for the network must be between lower and upper.
  }

}
}

\section{Constraint Descriptions}{

\itemize{
\item{\code{boundedDegree(lower,upper)}  (order-independent)  (undirected)}{ 
Adds a constraint that the degrees for the network must be between lower and upper.
  }

}
}

