% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{lologGmm}
\alias{lologGmm}
\title{Fits a LOLOG model via Generalized Method of Moments}
\usage{
lologGmm(formula, auxFormula = NULL, theta = NULL, nsamp = 1000,
  includeOrderIndependent = TRUE, weights = "full", tol = 0.1,
  nHalfSteps = 10, maxIter = 100, minIter = 2, startingStepSize = 0.1,
  maxStepSize = 0.5, cluster = NULL, verbose = TRUE)
}
\arguments{
\item{formula}{A lolog formula for the sufficient statistics}

\item{auxFormula}{A lolog formula of statistics to use for moment matching}

\item{theta}{Initial parameters values}

\item{nsamp}{The number of sample neteworks to draw at each iteration}

\item{includeOrderIndependent}{If true, all order independent terms in formula are used for moment matching.}

\item{weights}{The type of weights to use in the GMM objective. Either 'full' for the inverse of the full covariance matrix or 'diagnoal' for the inverse of the diagonal of the covariance matrix.}

\item{tol}{The Hotteling's T^2 p-value tolerance for convergance for the transformed moment conditions.}

\item{nHalfSteps}{The maximum number of half steps to take when the objective is not improved in an interation.}

\item{maxIter}{The maximum number of iterations.}

\item{minIter}{The minimum number of iterations.}

\item{startingStepSize}{The starting dampening of the parameter update.}

\item{maxStepSize}{The largest allowed value for dampening.}

\item{cluster}{A parallel cluster to use for graph simulation.}

\item{verbose}{Level of verbosity 0-2}
}
\value{
An object of class 'lolog'
}
\description{
Fits a LOLOG model via Generalized Method of Moments
}

