% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{lologVariational}
\alias{lologVariational}
\title{Fits a latent ordered network model using maximum likelihood}
\usage{
lologVariational(formula, nReplicates = 5L, downsampleRate = NULL,
  targetFrameSize = 5e+05)
}
\arguments{
\item{formula}{A lolog formula}

\item{nReplicates}{An integer controling how many dyad ordering to perform.}

\item{downsampleRate}{Controls what proportion of dyads in each ordering should be dropped.}

\item{targetFrameSize}{Sets downsampleRate so that the model frame for the logistic regression will have on average this amount of observations.}
}
\value{
An object of class 'lologVariationalFit'.
}
\description{
Fits a latent ordered network model using maximum likelihood
}
