% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{lologVariationalFit}
\alias{lologVariationalFit}
\title{Fits a latent ordered network model using maximum likelihood}
\usage{
lologVariationalFit(formula, order = NULL, nReplicates = 5L,
  downsampleRate = NULL, targetFrameSize = 5e+05)
}
\description{
Fits a latent ordered network model using maximum likelihood
}

