\name{MissingErnmModel}
\alias{MissingErnmModel}
\title{creates an ERNM likelihood model}
\usage{
  MissingErnmModel(observedSampler, unobservedSampler, ...)
}
\arguments{
  \item{observedSampler}{a sampler}

  \item{unobservedSampler}{a sampler conditional upon the
  observed values}

  \item{...}{additional parameters for the log likelihood}
}
\description{
  creates an ERNM likelihood model
}

