% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryNet.R
\name{plot.Rcpp_DirectedNet}
\alias{plot.Rcpp_DirectedNet}
\title{plot an DirectedNet object}
\usage{
\method{plot}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\description{
plot an DirectedNet object
}
