% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofit.R
\name{print.gofit}
\alias{print.gofit}
\title{prints a gofit object}
\usage{
\method{print}{gofit}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{passed to print.data.frame}
}
\description{
prints a gofit object
}
