% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{print.lolog}
\alias{print.lolog}
\title{Print a `lolog` object}
\usage{
\method{print}{lolog}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters (unused)}
}
\description{
Print a `lolog` object
}
