% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolog-variational.R
\name{print.lologVariationalFit}
\alias{print.lologVariationalFit}
\title{Print of a lologVariationalFit object}
\usage{
\method{print}{lologVariationalFit}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters (unused)}
}
\description{
Print of a lologVariationalFit object
}
