% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{simulate.lolog}
\alias{simulate.lolog}
\title{Generates BinaryNetworks from a fit lolog object}
\usage{
\method{simulate}{lolog}(object, nsim = 1, convert = FALSE, ...)
}
\arguments{
\item{object}{The lolog object}

\item{nsim}{The number of simulated networks}

\item{convert}{convert to a network object#'}
}
\description{
Generates BinaryNetworks from a fit lolog object
}
