% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{summary.lolog}
\alias{summary.lolog}
\title{Summary of a lolog object}
\usage{
\method{summary}{lolog}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{additional parameters (unused)}
}
\description{
Summary of a lolog object
}

