% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{ukFaculty}
\alias{ukFaculty}
\title{Friendship network of a UK university faculty}
\source{
The data set was origenally reported by Nepusz et. al. (2008) and was
subsequently processed and included by the igraphdata package. We have simply
converted their network from an igraph to a network object.
}
\usage{
data(ukFaculty)
}
\description{
The personal friendship network of a faculty of a UK university, 
consisting of 81 vertices (individuals) and 817 directed and weighted connections. 
The school affiliation of each individual is stored as a vertex attribute.
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors (Nepusz T., Petroczi A., Negyessy L., Bazso F. 2008)
should be cited, along with this \code{R} package.
}

\references{
Nepusz T., Petroczi A., Negyessy L., Bazso F.: Fuzzy communities and the 
concept of bridgeness in complex networks. Physical Review E 77:016107, 2008.
}
\keyword{datasets}
