\name{as.BinaryNet}
\alias{as.BinaryNet}
\title{convert and network to either an UndirectedNet or DirectedNet object}
\usage{
  as.BinaryNet(x, ...)
}
\arguments{
  \item{x}{the object}

  \item{...}{unused}
}
\description{
  convert and network to either an UndirectedNet or
  DirectedNet object
}

