% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-net.R
\name{as.network.Rcpp_DirectedNet}
\alias{as.network.Rcpp_DirectedNet}
\title{Convert a DirectedNet to a network object}
\usage{
\method{as.network}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
A network object
}
\description{
Convert a DirectedNet to a network object
}
\examples{
el <- matrix(c(1,2),ncol=2)

#make an UndirectedNet with one edge and 5 nodes
net <- new(UndirectedNet, el, 5L)

nw <- as.network(net)
nw
}
\seealso{
\code{\link{DirectedNet}}
}
