% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-net.R
\name{as.network.Rcpp_UndirectedNet}
\alias{as.network.Rcpp_UndirectedNet}
\title{Convert a UndirectedNet to a network object}
\usage{
\method{as.network}{Rcpp_UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{unused}
}
\value{
A network object
}
\description{
Convert a UndirectedNet to a network object
}
\examples{
el <- matrix(c(1,2),ncol=2)

#make an UndirectedNet with one edge and 5 nodes
net <- new(UndirectedNet, el, 5L)
net[1:5,1:5]

nw <- as.network(net)
nw
}
\seealso{
\code{\link{UndirectedNet}}
}
