% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-statistics.R
\name{calculateStatistics}
\alias{calculateStatistics}
\title{Calculate network statistics from a formula}
\usage{
calculateStatistics(formula)
}
\arguments{
\item{formula}{A lolog formula (See \code{\link{lolog}}).}
}
\description{
Calculate network statistics from a formula
}
\examples{
data(ukFaculty)
calculateStatistics(ukFaculty ~ edges + mutual + triangles)
}
