% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-model.R
\name{createCppModel}
\alias{createCppModel}
\title{Creates a model}
\usage{
createCppModel(formula, cloneNet = TRUE, theta = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{cloneNet}{create a deep copy of the network within the model object}

\item{theta}{the model parameters.}
}
\description{
Creates a model
}
\details{
Creates a C++ Model object. In general this isn't needed by most users of the
package.
}
\examples{
data(ukFaculty)
model <- createCppModel(ukFaculty ~ edges)
model$calculate()
model$statistics()
}
