% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolog.R
\name{createCppModel}
\alias{createCppModel}
\title{creates a model}
\usage{
createCppModel(formula, cloneNet = TRUE, theta = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{cloneNet}{create a deep copy of the network within the model object}

\item{theta}{the model parameters.}
}
\description{
creates a model
}

