\name{createCppModel}
\alias{createCppModel}
\title{creates a model}
\usage{
  createCppModel(formula, ignoreMnar = TRUE,
    cloneNet = TRUE)
}
\arguments{
  \item{formula}{the model formula}

  \item{ignoreMnar}{ignore missing not at random offsets}

  \item{cloneNet}{should the network be cloned}
}
\description{
  creates a model
}

