% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{createLatentOrderLikelihood}
\alias{createLatentOrderLikelihood}
\title{Creates a probability model for a latent ordered network model}
\usage{
createLatentOrderLikelihood(formula, theta = NULL)
}
\arguments{
\item{formula}{A LOLOG formula}

\item{theta}{Parameter values.}
}
\value{
An Rcpp object representing the likeilhood model
}
\description{
Creates a probability model for a latent ordered network model
}

