% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-net.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,Rcpp_DirectedNet-method}
\alias{[,Rcpp_DirectedNet,ANY,ANY,ANY-method}
\alias{\S4method{[}{Rcpp_DirectedNet,ANY,ANY,ANY}}
\alias{[}
\alias{[,Rcpp_UndirectedNet-method}
\alias{[,Rcpp_UndirectedNet,ANY,ANY,ANY-method}
\alias{\S4method{[}{Rcpp_UndirectedNet,ANY,ANY,ANY}}
\alias{[<-}
\alias{[<-,Rcpp_DirectedNet-method}
\alias{[<-,Rcpp_DirectedNet,ANY,ANY,ANY-method}
\alias{\S4method{[<-}{Rcpp_DirectedNet,ANY,ANY,ANY}}
\alias{[<-}
\alias{[<-,Rcpp_UndirectedNet-method}
\alias{[<-,Rcpp_UndirectedNet,ANY,ANY,ANY-method}
\alias{\S4method{[<-}{Rcpp_UndirectedNet,ANY,ANY,ANY}}
\title{indexing}
\usage{
\S4method{[}{Rcpp_DirectedNet,ANY,ANY,ANY}(x, i, j, ..., maskMissing = TRUE,
  drop = TRUE)

\S4method{[}{Rcpp_UndirectedNet,ANY,ANY,ANY}(x, i, j, ..., maskMissing = TRUE,
  drop = TRUE)

\S4method{[}{Rcpp_DirectedNet,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{Rcpp_UndirectedNet,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{object}

\item{i}{indices}

\item{j}{indices}

\item{...}{unused}

\item{maskMissing}{should missing values be masked by NA}

\item{drop}{unused}

\item{value}{values to assign}
}
\description{
indexing

indexing

indexing

indexing
}
\examples{
data(ukFaculty)
net <- as.BinaryNet(ukFaculty)


#dyad Extraction
net[1:2,1:5]
net$outNeighbors(c(1,2,3))

#dyad assignment
net[1,1:5] <- rep(NA,5)
net[1:2,1:5]
net[1:2,1:5,maskMissing=FALSE] #remove the mask over missing values and see
#nothing was really changed

#node variables
net$variableNames()
net[["Group"]]
net[["rnorm"]] <- rnorm(net$size())
net[["rnorm"]]
}
