\docType{methods}
\name{[}
\alias{[}
\alias{[,Rcpp_DirectedNet-method}
\alias{[,Rcpp_UndirectedNet-method}
\alias{[<-}
\alias{[<-,Rcpp_DirectedNet-method}
\alias{[<-,Rcpp_UndirectedNet-method}
\title{indexing}
\arguments{
  \item{x}{object}

  \item{i}{indices}

  \item{j}{indices}

  \item{...}{unused}

  \item{maskMissing}{should missing values be masked by NA}

  \item{drop}{unused}

  \item{x}{object}

  \item{i}{indices}

  \item{j}{indices}

  \item{...}{unused}

  \item{maskMissing}{should missing values be masked by NA}

  \item{drop}{unused}

  \item{x}{object}

  \item{i}{indices}

  \item{j}{indices}

  \item{...}{unused}

  \item{maskMissing}{should missing values be masked by NA}

  \item{value}{values to assign}

  \item{x}{object}

  \item{i}{indices}

  \item{j}{indices}

  \item{...}{unused}

  \item{maskMissing}{should missing values be masked by NA}

  \item{value}{values to assign}
}
\description{
  indexing

  indexing

  indexing

  indexing
}

