% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofit.R
\name{gofit}
\alias{gofit}
\title{Conduct goodness of fit diagnostics}
\usage{
gofit(object, ...)
}
\arguments{
\item{object}{the object to evaluate}

\item{...}{additional parameters}
}
\description{
Conduct goodness of fit diagnostics
}
\details{
see \code{\link{gofit.lolog}}
}
