% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{inlineLologPlugin}
\alias{inlineLologPlugin}
\title{An lolog plug-in for easy C++ prototyping and access}
\usage{
inlineLologPlugin(...)
}
\arguments{
\item{...}{plug-in arguments}
}
\description{
An lolog plug-in for easy C++ prototyping and access
}
\examples{
library(inline)
registerPlugin("lolog",inlineLologPlugin)
src <- "
	Rcpp::IntegerMatrix tmp(0,2);
	lolog::BinaryNet<lolog::Directed> net(tmp,Rcpp::as<int>(n));
	return net;
	"
emptyNetwork <- cxxfunction(signature(n="integer"), src, plugin="lolog")
net <- emptyNetwork(10)
net[1:10,1:10]
}

