% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{plot.gofit}
\alias{plot.gofit}
\title{Plots a gofit object}
\usage{
\method{plot}{gofit}(x, y, type = c("line", "box"), normalize = FALSE,
  line.alpha = 0.02, line.size = 1, ...)
}
\arguments{
\item{x}{the gofit object}

\item{y}{unused}

\item{type}{type of plot, boxplot or lineplot}

\item{normalize}{If true, netwrok statistics are normalized by subtracting off the observed statistics and scaling by the standard deviation.}

\item{line.alpha}{The transparancy of the simulated statistics lines}

\item{line.size}{The width of the lines}

\item{...}{passed to either boxplot or geom_line}
}
\description{
Plots a gofit object
}
\examples{
data(ukFaculty)
# A dyad independent model
fitind <- lolog(ukFaculty ~ edges() + nodeMatch("GroupC") + nodeCov("GroupC"))
summary(fitind)

# Check gof on degree distribution (bad!)
gind <- gofit(fitind, ukFaculty ~ degree(0:50))
plot(gind)
plot(gind, type="box")
}
