% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-net.R
\name{plot.Rcpp_DirectedNet}
\alias{plot.Rcpp_DirectedNet}
\title{plot an DirectedNet object}
\usage{
\method{plot}{Rcpp_DirectedNet}(x, ...)
}
\arguments{
\item{x}{the Rcpp_DirectedNet object}

\item{...}{additional parameters for plot.network}
}
\description{
plot an DirectedNet object
}
\details{
This is a thin wrapper around \code{\link{plot.network}}.
}
\examples{
data(ukFaculty)
net <- as.BinaryNet(ukFaculty)
plot(net, vertex.col=net[["Group"]]+1)
}
