% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-net.R
\name{plot.Rcpp_UndirectedNet}
\alias{plot.Rcpp_UndirectedNet}
\title{Plot an UndirectedNet object}
\usage{
\method{plot}{Rcpp_UndirectedNet}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters for plot.network}
}
\description{
Plot an UndirectedNet object
}
\details{
This is a thin wrapper around \code{\link{plot.network}}.
}
\examples{
el <- matrix(c(1,2),ncol=2)
net <- new(UndirectedNet, el, 5L)
net[1,5] <- 1
net[2,5] <- 1
plot(net)
}
