% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-order.R
\name{print.lolog}
\alias{print.lolog}
\title{Print of a lolog object}
\usage{
\method{print}{lologVariationalFit}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional parameters (unused)}
}
\description{
Print of a lolog object
}

