% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{registerDirectedStatistic}
\alias{registerDirectedStatistic}
\alias{registerUndirectedStatistic}
\alias{registerDirectedOffset}
\alias{registerUndirectedOffset}
\title{Register Statistics}
\usage{
registerDirectedStatistic
}
\description{
Register Statistics
}
